refactor: moved Price component
